RABSW-568 Use taints and tolerations for Rabbit nodes
Add a toleration to the nnf-sos node-manager daemonset.

In the KIND playground, taint the kind-worker* nodes as rabbit nodes.  This
means we have to move the nnf-controller-manager and dws-operator somplace
else.  We'll use the kind-control-plane node as a generic worker by removing
the default taint placed there by KIND, then label it to run the
controller-manager and dws-operator.

Update the kind-push target in the Makefile to push the nnf-sos container
to all nodes.