RABSW-801 Support for external Lustre MGT (#130)
Allow an external MGS NID to be specified in the nnf-ruleset.
Add a field to NnfStorge to take the external MGS NID.

Nnf_workflow_controller copies the external MGS NID from the #DW line into
the NnfStorage resource.

Directivebreakdown_controller does not create an "mgt" allocation when an
external MGS is specified.

Nnf_storage_controller copies the external MGS NID into the NnfNodeStorage, in
place of the SOS-built MGS.

Add support for the external MGS NID to the wfrLustre and
fluxPatchServersWorkflowSetup.sh tools. Getopts is added to wfrLustre to
make this easier. The wfrLustre tool is rewritten to
not need the kubectl proxy. Playground.sh is updated to match these changes.