RABSW-820: Add DeviceReference to all local ClientMounts (#138)
This commit fills in the DeviceReference information in the compute node ClientMount
resources for XFS and GFS2 file systems. This information isn't needed to do the actual
mount on the compute node, but it allows the data movement code to know which allocation
the ClientMount is mouning. This allows it to find the corresponding rabbit mount that
mounts the same allocation.

Co-authored-by: Matt Richerson <mattr@cray.com>