RABSW-855 Fix the project domain in the nnf-sos repo. (#153)
The PROJECT file in nnf-sos has a domain of "cray.com", but it should be
"cray.hpe.com".  That affects how operator-sdk creates resources and webhooks
and fills in apiGroups and kubebuilder directives.

I suppose people have been manually fixing up the operator-sdk generated output
to get things to work, and I hit this while trying to figure out why I couldn't
get a new webhook to run in nnfsos.

Yet, there are inconsistencies throughout the config/ dir that don't match
kubebuilder directives in other areas, making me wonder if we're actually using
everything that we think we're using.