RABSW-862 Validate NnfStorageProfile in proposal state. (#159)
To complete proposal state, verify that a named storage profile can be found or
that a default storage profile can be found.  When looking for a default
profile, verify that only one exists. Proposal state will fail if the named
profile cannot be found, or a default cannot be found, or there is more than
one default found.

Add a default profile to playground.sh.