RABSW-863 Use NnfStorageProfile when creating NnfStorages (#166)
The directivebreakdown_controller merges the Lustre part of the storage profile
with the lustre "combined_mgtmdt" or "external_mgs" directive arguments to
know which allocations to specify for a given jobdw or create_persisten DW
line.

A new file nnfstorageprofile_helpers.go has routines to make it easy to
work with storage profiles.  This is where we encapsulate the logic that
defines how to merge lustre "combined_mgtmdt" and "external_mgs" directive args
with the contents of a profile.

The findProfileToUse() function is re-worked and moved out of the
nnf_workflow_controller and into nnfstorageprofile_helpers.go so it can be used
by the nnf_workflow and directivebreakdown controllers.

Set a NNF_STORAGE_PROFILE_NAMESPACE env var in the nnf-controller-manager to
indicate the namespace of the storage profiles during a real run versus during
unit testing.