RABSW-904 Create immutable copies of NnfStorageProfiles that are in use (#178)
During proposal state create a copy of the NnfStorageProfile that is being used
for the storage. This copy of the resource will be used through the setup
state.  Initially it is owned by the workflow, and when the NnfStorage is
created that takes over as the owner; so the profile that was used to create
the storage will stick to the storage.

These immutable copies are called "pinned".  The NnfStorageProfile webhook
allows metadata updates on them, but not Data updates. They live in the same
namespace as the Workflow, DirectiveBreakdown, and NnfStorage so that the
ownerRef works for garbage collection.

There are several places where names are generated for DirectiveBreakdowns,
Servers, NnfStorages, PersistentStorageInstances.  These have been consolidated
in helper functions in nnf_workflow_controller_helpers.go.  These are used
to name the pinned NnfStorageProfile resources, so they match up with the
NnfStorage whether or not it is persistent.