RABSW-927: Include compute breakdown for persistentdw
This commit updates sos to use the new layout for the directiveBreakdown that
has a section for storage allocations and a section for compute allocations.
The compute information is filled in with a location constraint for specifying
how to choose compute nodes based on the storage they need access to. The
location constraint provides an ObjectReference to the Servers resource where
the storage is defined, and information on how that storage can be accessed
(physical vs. network).

I also changed the owner reference on the PersistentStorageInstance created by
the directiveBreakdown controller so that it's owned by the DirectiveBreakdown
through the Setup phase. This ensures that it's cleaned up if anything fails
while it's creating.

There were lots of code changes that were a result of the changed DirectiveBreakdown
fields and having a DirectiveBreakdown for persistentdw now.

Signed-off-by: Matt Richerson <mattr@cray.com>