RABSW-968: Verify NnfNodeStorage count when aggregating status
The NnfNodeStorage status aggregator uses a List() operation to find all the
NnfNodeStorage resources that are part of an allocation set. Depending on the
state of the client cache, we may miss some of the resources. This commit adds
a check to make sure we've found the expected number of resources.

Signed-off-by: Matt Richerson <mattr@cray.com>