RABSW-970: Process directives individually (#25)
* RABSW-970: Process directives individually

This commit changes the layout of the workflow controller to allow finer grained control of
which directives are processed, and how the success and failure of that directive is reported
in the workflow. It changes the "handler" functions to process a single directive specified
by its index in the workflow. This is different from before since the "handler" functions would
process all the directives in the workflow before marking them as successful.

The handler functions were changed to have a separate "start" and "finish" handler for each
workflow state (e.g., startProposalState and finishProposalState). This makes it easy to code
for adding parallelism when there are multiple directives. The "start" function creates all the
child objects, and the "finish" function waits for the child resource to become ready.

Which directives to process was standardized. Before we were doing a mixture of parsing the directive
list, walking the directive breakdowns, and looking at the driver status array in the workflow.
Everything is now based on the information in the driver status array of the workflow. This has
all the information of which directives are interested in which workflow states. It also allows us
to skip over directives that have already completed the current state.

Signed-off-by: Matt Richerson <mattr@cray.com>

* RABSW-970: Address review comments

- Use helper function to set directiveIndexLabel.
- Use indexedResourceName() for NnfAccess children
- Use g Gomega in test

Signed-off-by: Matt Richerson <mattr@cray.com>

* Increase EVENTUALLY timeout

Signed-off-by: Matt Richerson <mattr@cray.com>

* RABSW-970: Add failfast for tests

Signed-off-by: Matt Richerson <mattr@cray.com>

* RABSW-970: Fix typos in comments

Signed-off-by: Matt Richerson <mattr@cray.com>

* RABSW-970: Log error message

Signed-off-by: Matt Richerson <mattr@cray.com>

Co-authored-by: Matt Richerson <mattr@cray.com>