RABSW-985: Prefix webhook resource names for unit tests (#33)
* RABSW-985: Prefix webhook resource names for unit tests

This commit fixes an issue where the DWS workflow validation webhook was
colliding with the NNF storage profile validation webhook. This was because
the ValidatingWebhookConfiguration resource generated by the controller-gen
tool has the same name between the projects.

I added a script to prefix the names of the ValidatingWebhookConfiguration
and Service resource with "nnf-". The script makes a copy of the original
yaml files and passes the new location to the test code through an environment
variable.

Signed-off-by: Matt Richerson <mattr@cray.com>

* move export

Signed-off-by: Matt Richerson <mattr@cray.com>

Co-authored-by: Matt Richerson <mattr@cray.com>