RABSW-999: Use WorkflowError in workflow controller (#43)
* RABSW-999: Use WorkflowError in workflow controller

This commit uses the new WorkflowError type in the workflow controller.
This error type does a few things:
- Gives a user facing message about the error
- Specifies whether the error is likely recoverable
- Wraps an implementation level error

If the error is likely unrecoverable, then the workflow status is set to
"error". Otherwise the status is left as "running". In both cases, the
workflow message is set with the user facing message from the WorkflowError.

The message from the implementation level error is placed into a new field,
"error", in the driver status section.

Signed-off-by: Matt Richerson <mattr@cray.com>

* Review comments

Signed-off-by: Matt Richerson <mattr@cray.com>

* revendor

Signed-off-by: Matt Richerson <mattr@cray.com>

Co-authored-by: Matt Richerson <mattr@cray.com>