Added SerialNumber, FirmwareVersion, and Slot to NnfDriveStatus (#73)
Each new field is sourced from the `StorageController`'s corresponding
field.

In the case of `slot`, it is a concatenation of the
`StorageController`'s `Location` field. The `SeviceLabel` and
`LocationOrdinalValue` are put together. These two fields are sourced
from the StorageController's config file - see nnf-ec for more details.

For example, with a label of `J` and ordinal value of `12`, the
resulting `slot` is `J12`.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>