Allow NnfStorageProfile to pass args for mkfs, lvm, lustre
The NnfStorageProfile has been changed to take whole commandlines for volume
and filesystem creation.

The placeholder_nnfstorageprofile.yaml default profile has been updated
with commandlines which create the basic volumes and filesystems that EC
was creating before this switch to passing in whole commandlines. This profile
continues to be the default profile that is installed by nnf-deploy.

The NnfStorage and NnfNodeStorage resources have labels that name the pinned
profile that is being used with the workflow.

The nnf_node_storage_controller chooses appropriate commandlines from the
profile when constructing the OEM object for EC.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>