RABSW-1083: Change group and owner of storage
This commit passes the group ID and owner ID fields down to the NnfNodeStorage so it can be
used to set the owner/group of a file system. For XFS and gfs2, this is done through nnf-ec.
For Lustre file systems, all Lustre targets have to complete before the owner/group can be
set. The Lustre client mounts the file system on the Rabbit node hosting OST 0, and the owner/group
are set from there. This is coordinated through a "SetOwnerGroup" flag in the NnfNodeStorage
that is set once all the Lustre targets are ready.

Signed-off-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>