RABSW-1095: Unlock NnfStorage when NnfAccess is unmounted (#105)
* RABSW-1095: Unlock NnfStorage when NnfAccess is unmounted

The NnfStorage is locked for XFS file systems when an NnfAccess has the file system
mounted. This prevents the file system from being mounted in two places at once which
could lead to corruption issues. The lock was only removed when the NnfAccess was
deleted, but it needs to be removed when the NnfAccess is unmounted.

Signed-off-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>

* Review comments

Signed-off-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>

Signed-off-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>
Co-authored-by: Matt Richerson <matthew.richerson@Matts-MacBook-Pro.local>