RABSW-1107: Add container directive validation (#119)
These changes allow a user to create a container workflow and have it go ready in the `Proposal` state. It validates that:
- the supplied container profile exists
- if any non-optional storage in the container profile exists, ensure it is present in the supplied arguments to the container directive
- any supplied storage argument in the container directive must exist in the container profile

In order to support validation of the `container` directive, the `NnfStorageProfile` pinning and default functionality has been duplicated and refactored for `NnfContainerProfile`. Ideally in the future, we combine the two.

The following directives were used to test this container workflow:
```
#DW jobdw name=blake-gfs2 type=gfs2 capacity=50GB
#DW persistentdw name=blake-persistent
#DW container name=blake-container profile=nnfcontainerprofile-sample \
      JOB_DW_foo-local-storage=blake-gfs2 \
      PERSISTENT_DW_foo-persistent-storage=blake-persistent
```

This requires that a persistent storage instance is present prior to creating the container workflow. This means that you need a workflow with a `create_persistent` directive prior to creating a container.

Work still needs to be done on the directive breakdown controller.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>