RABSW-1124: Change NnfAccess TeardownState for servers (#136)
* RABSW-1124: Change NnfAccess TeardownState for servers

The data movement code was mounting and unmounting the Rabbit nodes during the DataIn
and DataOut phases of the workflow. A stale workflow resource in the client cache could
cause the NnfAccess to be re-mounted after it had already been unmounted. This commit
changes the NnfAccess Teardown state logic to do the unmounts in PreRun and Teardown
instead of DataIn and DataOut.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

* review comments

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>

---------

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>