RABSW-1129: Shorten LVM names
The LV and VG names were too long and caused an error during the lvcreate. This commit
changes the VG name to use a truncated version of the file share ID which includes the
workflow name/namespace, directive index, and allocation index. This string is combined
with the UUID of the workflow. The LV name was changed to be "lv" for all logical volumes
since there is only ever the single LV in each VG.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>