Use the live k8s client object in suite_test.go. (#175)
In the kubebuilder book:
https://book.kubebuilder.io/cronjob-tutorial/writing-tests.html
It explains that we should be using the "live" k8s client rather than the one
from the manager:

"Note that we set up both a “live” k8s client and a separate client from the
manager. This is because when making assertions in tests, you generally want to
assert against the live state of the API server. If you use the client from the
manager (k8sManager.GetClient), you’d end up asserting against the contents of
the cache instead, which is slower and can introduce flakiness into your
tests."