Add support for extra dcp and dryrun options in NnfDataMovementSpec (#180)
In order to support per-DM configuration options, we need to add some
options to the spec. These values will override/supplement the existing
data movement configuration that is defined in the nnf-dm-config
ConfigMap.

In this case, LLNL has a need to add extra dcp options for a given data
movement request. This will be done via the Copy Offload API.

For debugging purposes, the dryrun option has also been added to fake
out data movement.

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>