Restrict the name of the storage
Restrict the name of the storage to something that can also be used as the
name of a k8s resource.  This means the name must be made up of lowercase
letters, numbers, and dashes.

When using that name in a DW_JOB_ or DW_PERSISTENT_ environment variable, swap
the dashes for underscores in the storage name.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>