Containers: Add specific linux capabilities for MPI workers
On LLNL systems, the host's default capabilities are different than HPE
systems due to the difference in container runtime and also
configuration of those runtimes.

This change drops all capabilities and then adds in the required ones to
ensure mpirun can access the worker pods.

Additionally, fix some bugs in the MPISpec of the NnfContainerProfile:
- Don't allow NnfContainerProfile Data to be empty, it's required
- Added check to ensure MPISpec is defined
- Added checks to ensure launcher/worker are defined in the MPISpec

Tested capabilities by:
- Drop ALL capabilities without the specific Adds (and failing)
- Then adding in the specific Adds (and passing)

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>