Containers: Add global lustre support (#213)
This adds `DW_GLOBAL_*` storage support to container workflows/profiles,
which are backed by the `LustreFilesystem` resource. This allows users
to mount global lustre fileystems into user containers.

An administrator must first edit the `LustreFilesystem` resource to add the
workflow's namespace (e.g. `default`) to the list of namespaces along
with a mode. The mode defaults to `ReadWriteMany` if not set. Doing so
will create a PVC that can be used to mount the fileystem.

A `DW_GLOBAL_*` storage must also be added to the NnfContainer Profile.
See `config/examples` and `config/samples` for more detail.

The value for the `DW_GLOBAL_*` parameter is the path of the lustre
filesystem, e.g.:

```
#DW jobdw type=gfs2 name=my-local-storage capacity=100GB
#DW container name=my-container-workflow profile=example-mpi \
    DW_JOB_foo_local_storage=my-local-storage \
    DW_GLOBAL_foo_global_lustre=/lus/sawbill
```

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>