Containers: Allow containers to open ports for communication with compute nodes
This change allows for container workflows to open ports. These ports
are opened on the host nodes (i.e. NNF nodes) where the containers are
running. This enables traffic from outside of the network through the IP
address of the NNF node and the port. An application on the compute node
can contact the container with <NNF_NODE_IP>:<PORT>.

The port number(s) can be retrieved via the NNF_CONTAINER_PORT
environment variable. This environment variable is available inside of
the containers. It is also provided to the Workflow so that Flux can
inform the application on the compute node of which port(s) to use. If
multiple ports are desired, the environment variable will provide a
comma separated list of port numbers.

Ports are requested via the NnfContainerProfile's `numPorts`. **A system
admin must enable the `Ports` port range in the `SystemConfiguration`
before ports can be requested**. If not, the NnfPortManager will not
allocate any ports.

More details:
- Enabled default NnfPortManager to manage port allocation
- Port allocation occurs in the Setup State
- Port de-allocation occurs in the Teardown State
- User Container Pods are now destroyed in the Teardown State prior to
  Port de-allocation
- Added `example-mpi-webserver` NnfContainerProfile to show use of
  envionrment variable with a simple webserver
- Added container teardown + port allocation to workflow deletion