github-46: Use ResourceError when returning errors
This commit uses the new ResourceError struct embedded in the status section of the DWS/nnf-sos
resources. When returning an error, use the NewResourceError() call to return a ResourceError and
fill it in with the correct information. This allows the end user and WLM to make informed decisions
about what to do when there's an error.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>