Added Port Cooldown to NnfPortManager
Once ports have been released, they now go into a Cooldown period. The
cooldown period can be set via the SystemConfiguration. This defaults to
60s and corresponds with the kernel's default TIME_WAIT.

Once the cooldown period expires, the port is then removed from the
allocation list - freeing it up for reuse.

The cooldown period is checked in the following situations:

1. When a new user requests a port
2. When a user has requested a port but there are none left (forcing a
   requeue)
3. When a user has released a port

For #2, the reconciler will requeue for the cooldown period in hopes
that a port has been freed in that time. If not, the requque will
continue until that request has been satisfied.

If these situations do not occur, the port will remain in Cooldown even
though the time has expired. Since there is no reason to reconcile, the
cooldown period will not be checked until a new port is
requested/released (or
the reconcile fires for another reason).

Signed-off-by: Blake Devcich <blake.devcich@hpe.com>