File system and block device management in nnf-sos
This commit moves the file system, LVM, and zpool management out of nnf-ec and
into nnf-sos. There are many parts to this change:

- Create a package for running the LVM commands (e.g., pvcreate)
- Create a package for managing block devices. This presents as the blockdevice interface.
Implementations include lvm, zpool, and mock.
- Create a package for managing file systems. This presents as the filesystem interface.
Implementations include simple (for xfs, gfs2 , and bind mounts), lustre, mock, and kind.
- Create an NnfNodeBlockStorage resource that interfaces with nnf-ec to manage the
storage pools and storage groups.
- Change NnfNodeStorage to use the blockdevice and filesystem interfaces to create and delete
file systems on the Rabbit.
- Change the ClientMount reconciler to also use the blockdevice and filesystem interfaces to
activate/mount and unmount/deactivate filesystems and blockdevices.
- Move the clientmount daemon from DWS to nnf-sos. The ClientMount reconciler is now shared
between compute and Rabbit nodes (run as clientmountd on the computes and inside nnf-node-manager
on the Rabbits), and it is NNF specific (not generic for DWS).
- Add command line arguments to the NnfStorageProfile for activating/deactivating single LVs
- Use a shared VG with locking for all LVM VGs since they're always shared between multiple nodes.
This causes LVM to enforce a single node activating a non-shared LV at a time. In the future we can
remove the internal locking nnf-sos does in the NnfAccess reconciler
- Use the ClientMount resource to mount the Lustre file system after creation to set the GID/UID
- Try to consistently use MgsAddress as the resource field for the MGS LNet Address
- Turn NnfNodeStorage Status section into a subresource

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>