Requirements for Prometheus
Add a label to the ServiceMonitor resource that will be common across all
DWS and NNF repos so we can easily tell Prometheus which ones it should select.

Copy the namespace value into config/prometheus so that part of config/ can
be deployed on its own.

Add a ClusterRoleBinding that can be used by a metrics reader process to
debug the metrics endpoint of the controller.  The ClusterRole is already
being installed, and the ServiceAccount is already installed.

Update the deploy.sh script to determine whether or not the ServiceMonitor
resource can be applied.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>