Use owner UID instead of workflow UID for VG name
This is needed for persistent local file systems. The workflow mounting the persistent
file system will have a different UID than the workflow that made the file system. Use
the owner UID instead (NnfStorage) since that UID will be consistent for the lifetime
of the persistent file system.

Also, don't set the file system permissions on ClientMounts in Kind. This will be enabled
later when file systems are better mocked out in the kind environment.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>