Add option for a single shared VG per Rabbit per workflow
This commit adds a new option, sharedVg, in the NnfStorageProfile for allocations backed by
an LVM logical volume. This option allows allocations on a Rabbit node that are part of the
same workflow to share a volume group. This reduces the overhead of creating allocations
and greatly speeds up the Setup and Teardown phases of the workflow.

This commit also changes the way locking is done with lvmlockd. Previously, when the LV was
not needed on a node, the LV was deactivated and the VG lockspace was stopped. Now, only the
LV is deactivated in most situations. The VG lockspace is only stopped when a node is completely
done with a VG and it won't be activated again.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>