Let the SystemConfiguration controller add taints and labels to rabbits
This controller is now responsible for adding the taints and labels to the
rabbit nodes.  It does this in three passes.

Pass 1 adds cray.nnf.node label and NoSchedule taint to each rabbit.  It skips
any rabbit that is already marked with the cray.nnf.node.cleared label, which
marks that rabbit as finished.  If any rabbit is modified in this pass then
it asks to re-enter the reconciler before going to the next pass.

Pass 2 adds the NoExecute taint to each rabbit.  Again, it skips any rabbit
that is already maked with the cray.nnf.node.cleared label.  This taint will
evict any pod that isn't supposed to be on a rabbit; unfortunately, it will
also evict everything else, but we'll open it for scheduling again in the
next pass.

Pass 3 adds the cray.nnf.node.cleared label and removes the NoExecute taint.
This is the final step for each rabbit.

In Pass 2 we ensure that all rabbits get the NoExecute taint together, except
those rabbits that are marked as already finished (with the
cray.nnf.node.cleared label).  This means that the pods we're trying to evict
will be squeezed out of the rabbits; by this time all rabbits will
have the NoSchedule taint from Pass 1, so we won't be playing whack-a-mole with
those pods.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>