Make sure device paths are populated in the NnfNodeBlockStorage
Don't return a nil Result after creating the StorageGroup. This causes the
NnfNodeBlockStorage to be marked as Ready before the /dev devices have
been added to the status section. The NnfNodeStorage reconciler won't
find any devices to add as LVM PVs, and it will fail trying to do a vgcreate
with no PVs.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>