Add NnfLustreMGT resource for tracking fsname
Create a new NnfLustreMGT resource that tracks used fsnames and erases unused
fsnames on internal MGTs. A NnfLustreMGT resource is created automatically whenever
a Lustre MGT is created. Lustre file systems that want to use the MGT can add a
claim to the NnfLustreMGT resource to request an fsname. The NnfLustreMGT controller
will find the next available fsname and assign it. When a Lustre file system is
torn down, removing the claim from the NnfLustreMGT resource will trigger the
NnfLustreMGT controller to run an `lctl` command on the Rabbit to erase the fsame
from the MGT so it can be reused.

For external MGTs, a NnfLustreMGT resource must be created to track fsnames. No
lctl commands are run on external MGTs, so the fsname will eventually wrap and
cause an error when creating a lustre file system. External MGTs should make use
of the FsNameStartReference field in the NnfLustreMGT resource to reference a config
map where the next available fsname can be stored. This configmap should not be
managed by argocd so that it persists across undeploys/deploys.

The NnfLustreMGT controllers run on both the Rabbits and a worker node. The controllers
running on the Rabbits handle NnfLustreMGT resources for internal MGTs where the lctl
command can be run. The controller running on the k8s worker handles NnfLustreMGTs
that describe external MGTs.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>