Don't check status when deleting storage pools
The allocation teardown code was checking whether a storage pool ID was listed in the
status section of the NnfNodeBlockStorage before trying to delete it. If there was
an issue saving the storage pool ID (e.g., conflict), then we can strand NVMe namespaces
on the Rabbit. Always try to delete storage pools for every known index and treat "not found"
errors as a success.

Also, fix an error with the NnfStorage deletion code where the child objects (NnfNodeStorage and
NnfNodeBlockStorage) don't get deleted if they currently have an error.

Signed-off-by: Matt Richerson <matthew.richerson@hpe.com>