Identify inability to open a switch, avoid nil dereference
In the case where we cannot open a /dev/switchtecX device, give a useful panic
message rather than just dereferencing a nil pointer.

Earlier in the log there will be an error stating "Error opening path", and
another stating "Failed to identify switch". The nil deference was happening
much later, so it was not obviously connected to those.

Signed-off-by: Dean Roehrich <dean.roehrich@hpe.com>